refactor Chart component to extract zoom logic into createChartZoom factory function, consolidate scaleData hash computations into hoisted variables scaleDataHashX and scaleDataHash, add visibleCandleDataHash and drawnShapeHistoryHash for reuse across dependency arrays, and extract hover status logic into useHoverStatus hook
